Made floating point clock initializers disable symbolic
