+++ .gitignore
@@ -5,3 +5,4 @@
 /.idea
 /result
 /libs
+/.vscode
